oracle_bridge/requirements.txt
flask>=2.3.0
web3>=6.0.0
python-dotenv>=0.21.0 